refactor out PathElement dependency cycle
